Move DataCoordinate container classes to new modules.
This starts a new _containers subpackage that will be extended to
include other butler primitives, like DimensionRecords and DatasetRefs.
