+++ python/lsst/daf/butler/core/__init__.py
@@ -33,3 +33,4 @@ from . import progress  # most symbols are only used by handler implementors
 # Only import ButlerLogRecords
 # ButlerLogRecords is the fundamental type stored in datastores.
 from .logging import ButlerLogRecords
+from ._containers import *
